+++ src/PsDac/InstallPackageCommand.cs
@@ -59,6 +59,24 @@ namespace PsDac
             }
         }
 
+        [Parameter()]
+        public SwitchParameter IgnoreWithNocheckOnForeignKeys
+        {
+            set { DeployOptions.IgnoreWithNocheckOnForeignKeys = value.IsPresent;  }
+        }
+
+        [Parameter()]
+        public SwitchParameter IgnoreWithNocheckOnCheckConstraints
+        {
+            set { DeployOptions.IgnoreWithNocheckOnCheckConstraints  = value.IsPresent;  }
+        }
+
+        [Parameter()]
+        public SwitchParameter SkipNewConstraintValidation
+        {
+            set { DeployOptions.ScriptNewConstraintValidation = !value.IsPresent; }
+        }
+
         #region Timeouts
 
         [Parameter()]