+++ CHANGELOG.md
@@ -7,11 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.7.0] - 2024-08-13
+
 ### Changed
 
 - Added support for views to `Get-DacColumn` command.
 - Updated Microsoft.SqlServer.DacFx to 162.1.167.
 - Updated System.Management.Automation to 7.2.16 and PowerShell 7.2.16.
+
+### Added
+
+- Added `Install-Package` parameter for IgnoreWithNocheckOnCheckConstraints, IgnoreWithNocheckOnForeignKeys and ScriptNewConstraintValidation.
 
 ## [1.6.0] - 2023-08-24